requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.2.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
aiosqlite==0.19.0
strawberry-graphql[fastapi]>=0.200.0
jinja2==3.1.2
python-multipart==0.0.6
